+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `hide_inactive_statusline` option deprecated from `lualine` plugin
+- `javascriptTSTagAttribute` highlight removed
+- `TSTagAttribute` highlight added
+- yellow changed to the same version as onedark palette. monsonjeremy/onedark.nvim#19
+- Boolean set to orange as in VSC onedark theme. monsonjeremy/onedark.nvim#19
 
 ## [v0.0.3]- 11 Dec 2021